+++ include/flag_gems/utils.h
@@ -17,4 +17,9 @@ std::filesystem::path get_triton_src_path();
 std::filesystem::path get_flag_gems_src_path();
 int64_t next_power_of_2(int64_t n);
 bool broadcastable_to(at::IntArrayRef s1, at::IntArrayRef s2);
+std::tuple<at::Tensor, int64_t, int64_t> permute_reduction_axes_right(const at::Tensor &tensor,
+                                                                      int reduction_axis);
+std::tuple<at::Tensor, int64_t, int64_t> permute_reduction_axes_right(
+    const at::Tensor &tensor, at::OptionalIntArrayRef reduction_axes_opt);
+int cdiv(int a, int b);
 }  // namespace flag_gems::utils
