+++ lib/utils.cpp
@@ -8,7 +8,7 @@ std::filesystem::path get_path_of_this_library() {
   // there is no build system generator to take care of this.
   static const std::filesystem::path cached_path = []() {
     Dl_info dl_info;
-    if (dladdr(reinterpret_cast<void*>(&get_path_of_this_library), &dl_info) && dl_info.dli_fname) {
+    if (dladdr(reinterpret_cast<void *>(&get_path_of_this_library), &dl_info) && dl_info.dli_fname) {
       return std::filesystem::canonical(dl_info.dli_fname);  // Ensure absolute, resolved path
     } else {
       throw std::runtime_error("cannot get the path of libjit_utils.so");
@@ -34,7 +34,7 @@ std::filesystem::path get_triton_src_path() {
 
 std::filesystem::path get_flag_gems_src_path() {
   const static std::filesystem::path flag_gems_src_dir = []() {
-    const char* flag_gems_dir = std::getenv("FLAGGEMS_SOURCE_DIR");
+    const char *flag_gems_dir = std::getenv("FLAGGEMS_SOURCE_DIR");
     if (!flag_gems_dir) {
       throw std::runtime_error("Environment variable FLAGGEMS_SOURCE_DIR not set");
     }
@@ -79,5 +79,58 @@ bool broadcastable_to(at::IntArrayRef s1, at::IntArrayRef s2) {
   }
 
   return true;
+}
+std::tuple<at::Tensor, int64_t, int64_t> permute_reduction_axes_right(
+    const at::Tensor &tensor, at::OptionalIntArrayRef reduction_axes_opt) {
+  int64_t dim = tensor.dim();
+  c10::DimVector reduction_axes;
+
+  if (reduction_axes_opt.has_value()) {
+    reduction_axes = reduction_axes_opt.value().vec();
+  }
+
+  std::unordered_set<int64_t> reduction_set(reduction_axes.begin(), reduction_axes.end());
+
+  c10::DimVector left_axes, right_axes;
+  int64_t non_reduction_size = 1, reduction_size = 1;
+
+  for (int64_t i = 0; i < dim; ++i) {
+    if (reduction_set.count(i)) {
+      right_axes.push_back(i);
+      reduction_size *= tensor.size(i);
+    } else {
+      left_axes.push_back(i);
+      non_reduction_size *= tensor.size(i);
+    }
+  }
+
+  // Concatenate left and right axes to form the new permutation order
+  c10::DimVector permute_order = left_axes;
+  permute_order.insert(permute_order.end(), right_axes.begin(), right_axes.end());
+
+  return {tensor.permute(permute_order), non_reduction_size, reduction_size};
+}
+std::tuple<at::Tensor, int64_t, int64_t> permute_reduction_axes_right(const at::Tensor &tensor,
+                                                                      int reduction_axis) {
+  int64_t dim = tensor.dim();
+  c10::DimVector left_axes, right_axes;
+  int64_t non_reduction_size = 1, reduction_size = 1;
+
+  for (int64_t i = 0; i < dim; ++i) {
+    if (i == reduction_axis) {
+      right_axes.push_back(i);
+      reduction_size *= tensor.size(i);
+    } else {
+      left_axes.push_back(i);
+      non_reduction_size *= tensor.size(i);
+    }
+  }
+  c10::DimVector permute_order = left_axes;
+  permute_order.insert(permute_order.end(), right_axes.begin(), right_axes.end());
+
+  return {tensor.permute(permute_order), non_reduction_size, reduction_size};
+}
+int cdiv(int a, int b) {
+  return (a + b - 1) / b;
 }
 }  // namespace flag_gems::utils
